+++ tutorial/src/main/scala-0.22/tutorial/libMyCatsForExamples01To09/Id.scala
@@ -1,0 +1,3 @@
+package tutorial.libMyCatsForExamples01To09
+
+type Id[A] = A
