+++ tutorial/src/main/scala-0.22/tutorial/libMyCatsForExamples01To09/Reader.scala
@@ -1,4 +1,4 @@
-package tutorial.libMyCatsUntilExample09
+package tutorial.libMyCatsForExamples01To09
 
 import scala.annotation.infix
