--- tutorial/src/main/scala-0.22/tutorial/libMyCatsUntilExample09/Id.scala
@@ -1,3 +1,0 @@
-package tutorial.libMyCatsUntilExample09
-
-type Id[A] = A
